Packaged everything into smaller modules, have also changed task.html to updateTask.html
