+++ dopeticksApp/dopeticks/templates/account.html
@@ -1,6 +1,5 @@
 {% extends "layout.html" %}
 {% block content %}
-	<p> This is the User Account's page (previously the Profile page) </p>
 
 	<div class="content-section">
 	  <div class="media col-md-4">
