+++ dopeticksApp/dopeticks/templates/home.html
@@ -9,8 +9,8 @@
 				<div class="col-md-8 col-lg-6 col-xl-4 offset-xl-1">
 					<!--Just basic text here-->
 					<h2>Get your dopamine fix and manage your tasks with DopeTicks today!</h2>
-					<a href={{ url_for('register') }} class="btn btn-sm btn-outline-primary">Create New Account</a>
-					<a href={{ url_for('login') }} class="btn btn-sm btn-outline-secondary">Log In</a>
+					<a href={{ url_for('users.register') }} class="btn btn-sm btn-outline-primary">Create New Account</a>
+					<a href={{ url_for('users.login') }} class="btn btn-sm btn-outline-secondary">Log In</a>
 
 					
 				</div>
