+++ dopeticksApp/dopeticks/templates/layout.html
@@ -34,9 +34,9 @@
 		      <div class="collapse navbar-collapse" id="navbarToggle">
 		        <div class="navbar-nav mr-auto">
 		        	{% if current_user.is_authenticated %}
-		        		<a class="nav-item nav-link"  style="color:#424242;" href="{{ url_for('dashboard') }}">Dashboard</a>
+		        		<a class="nav-item nav-link"  style="color:#424242;" href="{{ url_for('main.dashboard') }}">Dashboard</a>
 		        	{% else %}
-		          	<a class="nav-item nav-link" style="color:#424242;" href="{{ url_for('home') }}">Home</a>
+		          	<a class="nav-item nav-link" style="color:#424242;" href="{{ url_for('main.home') }}">Home</a>
 		        	{% endif %}
 		          
 		        </div>
@@ -44,11 +44,11 @@
 		        <div class="navbar-nav">
 		        	<!-- # Using ginja2 conditional and using current_user.is_authenticated to check if user is logged in or not -->
 		        	{% if current_user.is_authenticated %}		
-		        		<a class="nav-item nav-link"  style="color:#424242;" href="{{ url_for('account') }}">{{ current_user.userFirstName }} {{ current_user.userLastName }}</a>
-		        		<a class="nav-item nav-link"  style="color:#424242;" href="{{ url_for('logout') }}">Logout</a>
+		        		<a class="nav-item nav-link"  style="color:#424242;" href="{{ url_for('users.account') }}">{{ current_user.userFirstName }} {{ current_user.userLastName }}</a>
+		        		<a class="nav-item nav-link"  style="color:#424242;" href="{{ url_for('users.logout') }}">Logout</a>
 		        	{% else %}
-			          <a class="nav-item nav-link"  style="color:#424242;" href="{{ url_for('login') }}">Login</a>
-			          <a class="nav-item nav-link"  style="color:#424242;" href="{{ url_for('register') }}">Register</a>
+			          <a class="nav-item nav-link"  style="color:#424242;" href="{{ url_for('users.login') }}">Login</a>
+			          <a class="nav-item nav-link"  style="color:#424242;" href="{{ url_for('users.register') }}">Register</a>
 		          {% endif %}
 		        </div>
 		      </div>
