+++ dopeticksApp/dopeticks/templates/login.html
@@ -55,7 +55,7 @@
 							<!--link to the registration page-->
 							<div class="text-center text-lg-start mt-4 pt-2">
 								<small class="small mt-2 pt-1 mb-0">
-									Want an account? <a class="link-primary" href="{{ url_for('login') }}">Sign up here</a>
+									Want an account? <a class="link-primary" href="{{ url_for('users.register') }}">Sign up here</a>
 								</small>
 							</div>
 					</form>
