+++ dopeticksApp/dopeticks/templates/register.html
@@ -67,7 +67,7 @@
 						</div>
 						<div class="text-center text-lg-start mt-4 pt-2">
 							<small class="small mt-2 pt-1 mb-0">
-							Already have an account? <a class="link-primary" href="{{ url_for('login') }}">Log in here</a>
+							Already have an account? <a class="link-primary" href="{{ url_for('users.login') }}">Log in here</a>
 							</small>
 						</div>
 					</form>
